pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bunkervm"
version = "0.4.0"
description = "Hardware-isolated Linux sandbox for AI agents — Firecracker MicroVM + MCP"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [{name = "Ashish"}]

keywords = ["mcp", "sandbox", "firecracker", "ai", "agent", "microvm", "isolation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Emulators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
langgraph = [
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
]
openai-agents = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.0.0",
]
crewai = [
    "crewai>=0.50.0",
    "python-dotenv>=1.0.0",
]
all = [
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.1.0",
    "openai-agents>=0.1.0",
    "crewai>=0.50.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bunkervm = "bunkervm.cli:main"

[project.urls]
Homepage = "https://github.com/ashishgituser/bunkervm"
Documentation = "https://github.com/ashishgituser/bunkervm#readme"
Issues = "https://github.com/ashishgituser/bunkervm/issues"

[tool.setuptools.packages.find]
include = ["bunkervm*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
